Add playGame function and logic
I added a playGame function to iteratively control the number of times
the playRound function is called. In this way, the user can play as many
times as they want against the computer using a for loop. Thus, the
logic of the playRound function is moved inside the playGame function.
Additionally, the scores and winner are displayed at the end of the
game.