Add button class and make playerSelection function
I first add defer to the script tag so that the script runs after the
DOM is loaded. Next, I added button classes for each selection buttons.
After this, I created a playerSelection function that will listen for
the button click events. Depending on the button clicked, the proper
alert message is displayed, which is used for testing the function. I
also temporarily disabled the playGame function.