Add images and CSS, modify HTML and JavaScript
I added images to serve as illustrations for the hand choices in the
game. I also added a stylesheet to prepare the website for design and
provide hide functionality.

I modified the HTML page to restructure its content and add initial
classes. This is then utilized by the JS script.

In the script, I built the logic of the game from the ground-up while
using the button's addEventlistener as input. I also utilized DOM
manipulation to provide functionality to the HTML content of the game.

It took a lot of trials and error but I managed to make it work. Go me!